refactor(memory): split migration diagnosis responsibilities
